Use size_t for index argument
This matches the OpenCoarrays public API, as well as the MPI_Aint type
used in MPI_Compare_and_swap and MPI_Fetch_and_op function signatures.

+++ src/mpi/mpi_caf.c
@@ -441,7 +441,7 @@ unimplemented_alloc_comps_message(const char * functionname);
 
 static void
 locking_atomic_op(MPI_Win win, int *value, int newval,
-                  int compare, int image_index, int index)
+                  int compare, int image_index, size_t index)
 {
   CAF_Win_lock(MPI_LOCK_EXCLUSIVE, image_index - 1, win);
   int ierr = MPI_Compare_and_swap(&newval, &compare,value, MPI_INT,
@@ -652,7 +652,7 @@ redo:
 }
 #endif
 
-void mutex_lock(MPI_Win win, int image_index, int index, int *stat,
+void mutex_lock(MPI_Win win, int image_index, size_t index, int *stat,
                 int *acquired_lock, char *errmsg, size_t errmsg_len)
 {
   const char msg[] = "Already locked";
@@ -737,7 +737,7 @@ stat_error:
 #endif // MPI_VERSION
 }
 
-void mutex_unlock(MPI_Win win, int image_index, int index, int *stat,
+void mutex_unlock(MPI_Win win, int image_index, size_t index, int *stat,
                   char* errmsg, size_t errmsg_len)
 {
   const char msg[] = "Variable is not locked";
